refactor: entity 파일 위치 common 안으로 변경
