+++ server/src/app.module.ts
@@ -6,11 +6,11 @@ import { TypeOrmModule } from "@nestjs/typeorm";
 import * as redisStore from "cache-manager-ioredis";
 import { AppController } from "./app.controller";
 import { AppService } from "./app.service";
-import { User } from "./entities/user.entity";
-import { Course } from "./entities/course.entity";
-import { Recruit } from "./entities/recruit.entity";
-import { UserRecruit } from "./entities/user_recruit.entity";
-import { HDong } from "./entities/h_dong.entity";
+import { User } from "./common/entities/user.entity";
+import { Course } from "./common/entities/course.entity";
+import { Recruit } from "./common/entities/recruit.entity";
+import { UserRecruit } from "./common/entities/user_recruit.entity";
+import { HDong } from "./common/entities/h_dong.entity";
 import { UserModule } from "./user/user.module";
 import { AuthModule } from "./auth/auth.module";
 import type { ClientOpts } from "redis";
