+++ server/src/common/repository/h_dong.repository.ts
@@ -1,5 +1,5 @@
 import { CustomRepository } from "src/common/typeorm/typeorm.decorator";
-import { HDong } from "src/entities/h_dong.entity";
+import { HDong } from "src/common/entities/h_dong.entity";
 import { Repository } from "typeorm";
 
 @CustomRepository(HDong)
