+++ server/src/course/course.module.ts
@@ -3,7 +3,7 @@ import { CourseService } from "./course.service";
 import { CourseController } from "./course.controller";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { TypeOrmCustomModule } from "src/common/typeorm/typeorm.module";
-import { Course } from "src/entities/course.entity";
+import { Course } from "src/common/entities/course.entity";
 import { CourseRepository } from "./course.repository";
 
 @Module({
