+++ server/src/course/dto/create-course.dto.ts
@@ -1,7 +1,7 @@
 import { IsNumber, IsNumberString, IsString } from "class-validator";
 import { isValidPath } from "src/common/decorator/path.validator";
 import { LatLng } from "src/common/type/lat-lng";
-import { Course } from "src/entities/course.entity";
+import { Course } from "src/common/entities/course.entity";
 
 export class CreateCourseDto {
     @IsString()
