+++ server/src/recruit/recruit.repository.ts
@@ -1,5 +1,5 @@
 import { CustomRepository } from "src/common/typeorm/typeorm.decorator";
-import { Recruit } from "src/entities/recruit.entity";
+import { Recruit } from "src/common/entities/recruit.entity";
 import { Repository } from "typeorm";
 import { RawRecruitData } from "src/common/type/raw-recruit-data";
 import { BadRequestException } from "@nestjs/common";
