+++ server/src/user/user.module.ts
@@ -3,7 +3,7 @@ import { UserService } from "./user.service";
 import { UserController } from "./user.controller";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { TypeOrmCustomModule } from "src/common/typeorm/typeorm.module";
-import { User } from "src/entities/user.entity";
+import { User } from "src/common/entities/user.entity";
 import { UserRepository } from "./user.repository";
 import { UserRecruitRepository } from "src/user_recruit.repository";
