+++ server/src/user/user.repository.ts
@@ -1,5 +1,5 @@
 import { CustomRepository } from "src/common/typeorm/typeorm.decorator";
-import { User } from "src/entities/user.entity";
+import { User } from "src/common/entities/user.entity";
 import { Repository } from "typeorm";
 
 @CustomRepository(User)
