+++ server/src/user_recruit.repository.ts
@@ -1,6 +1,6 @@
 import { CustomRepository } from "src/common/typeorm/typeorm.decorator";
 import { Repository } from "typeorm";
-import { UserRecruit } from "./entities/user_recruit.entity";
+import { UserRecruit } from "./common/entities/user_recruit.entity";
 
 @CustomRepository(UserRecruit)
 export class UserRecruitRepository extends Repository<UserRecruit> {
